// src/types.ts
export type BehaviorAction =
  | 'idle'
  | 'walking'
  | 'running'
  | 'sitting'
  | 'sleeping'
  | 'eating'
  | 'drinking'
  | 'playing'
  | 'peeing'
  | 'pooping'
  | 'speaking'
  | 'interacting'
  // New abilities (skeletal-clip overhaul)
  | 'wagging'
  | 'stretching'
  | 'shaking'
  | 'digging';
// ...
export type PetObjectKind =
  | 'dog_house'
  | 'food_bowl'
  | 'water_bowl'
  | 'ball'
  | 'bone'
  | 'bed'
  | 'hydrant'
  | 'chew_toy';